requirements.txt
requirements.txtrequests
beautifulsoup4
lxml